+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "guac-chat",
-  "version": "2.0.0-alpha.0",
+  "version": "2.0.0-alpha.1",
   "description": "Chat for guac",
-  "main": "src/index.js",
+  "main": "lib/src/server.js",
   "scripts": {
     "build": "rimraf dist && tsc -p tsconfig.build.json",
-    "dev": "tsc-watch -p tsconfig.json --onSuccess 'node ./lib/index.js'",
+    "dev": "tsc-watch -p tsconfig.json --onSuccess 'node ./lib/src/server.js'",
     "format": "prettier --write \"src/**/*.ts\"",
     "lint": "tslint -p tsconfig.json -c tslint.json",
     "typecheck": "tsc -p tsconfig.build.json --noEmit",
@@ -32,6 +32,7 @@
   },
   "dependencies": {
     "@types/escape-html": "^1.0.0",
+    "@types/express": "^4.17.11",
     "@types/flake-idgen": "^0.1.30",
     "@types/nconf": "^0.10.0",
     "@types/redis": "^2.8.28",
