+++ src/server.ts
@@ -1,0 +1,3 @@
+import {ChatServer} from './ChatServer';
+const app = new ChatServer().app;
+export default app;
