+++ tsconfig.build.json
@@ -1,5 +1,6 @@
 {
     "extends": "./tsconfig.json",
+    "include": ["src/**/*.ts"],
     "exclude": [
       "node_modules",
       "tests"
