+++ tsconfig.json
@@ -1,10 +1,11 @@
 {
 	"compilerOptions": {
 		"module": "commonjs",
-		"lib": [ "es2015", "es2020", "dom" ],
-		"target": "es2020",
+		"target": "es2017",
+		"declaration": true,
+		"moduleResolution": "node",
 		"sourceMap": true,
-		"outDir": "./lib",
+		"outDir": "lib",
 		"typeRoots": ["node_modules/@types", "src/types"],
 		"noUnusedLocals": true,
 		"noUnusedParameters": true,
@@ -13,8 +14,8 @@
 		"emitDecoratorMetadata": true,
 		"experimentalDecorators": true,
 		"resolveJsonModule": true,
-		"downlevelIteration": true,
 		"esModuleInterop": true
 	},
+	"include": ["src/**/*"],
 	"exclude": ["node_modules", "**/*.spec.ts"]
 }
